Add FileType exclusion
Added exclusions to the copy function, to ensure DLL and EXE files are excluded - these are typically not desired, and should not be included in a backup - use the Installed version!

Exclusions list can be edited anytime.
Ensured the Variables can be easily found and adjusted at the top if required.